Adding Arugment Completion for the TimeZone parameters (#3)
Alll parameters, -TimeZone, -ToTimeZone, and -FromTimeZone now have argument completers. This is only available in powershell version 5 and above. The code double-checks that it is v5 before attempting to load the argument completer powershell script. There is also a bugfix for the Convert-UtcTolocal when no time is passed in from the command line.